requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Login>=0.6,<1
Pillow>=10,<13
Werkzeug>=3,<4
gunicorn>=21,<24
